Move all code specific to 223 platform to common source set
